Ensure debounce on latest update

+++ duck-bot-nest/src/middlewares/debounce.middleware.ts
@@ -2,7 +2,9 @@ import { MiddlewareFn } from 'telegraf';
 import { SessionContext } from 'telegraf/typings/session';
 import { UpdateType } from 'telegraf/typings/telegram-types';
 
-const callbacks = new Map<string, NodeJS.Timeout>();
+type DebounceCallback = { update_id: number; timeout: NodeJS.Timeout };
+
+const callbacks = new Map<string, DebounceCallback>();
 
 export function telegrafDebounce<TSession extends Record<string, unknown>>(
   debounceTime = 1000,
@@ -16,28 +18,41 @@ export function telegrafDebounce<TSession extends Record<string, unknown>>(
     }
 
     const key = `${ctx.from.id}:${ctx.updateType}`;
+    const update_id = ctx.update.update_id;
+
+    // ensure debounce on latest update
+    if (!canModify(key, update_id)) {
+      console.log(key, 'received old update');
+      return;
+    }
 
     deleteCallback(key);
-    createCallback(key, next, debounceTime);
+    createCallback(key, update_id, next, debounceTime);
   };
 }
 
-function createCallback(key: string, fn: () => Promise<void>, debounceTime: number): void {
+function canModify(key: string, update_id: number): boolean {
+  return callbacks.get(key)?.update_id < update_id;
+}
+
+function createCallback(key: string, update_id: number, fn: () => Promise<void>, debounceTime: number): void {
   const timeOut = setTimeout(async () => {
     console.log(key, 'debounce');
     await fn();
     deleteCallback(key);
   }, debounceTime);
-  callbacks.set(key, timeOut);
+
+  callbacks.set(key, { update_id: update_id, timeout: timeOut });
   console.log(key, 'created');
 }
 
 function deleteCallback(key: string): void {
-  const timeOut = callbacks.get(key);
-  if (!timeOut) {
+  const debounceCallback = callbacks.get(key);
+  if (!debounceCallback) {
     return;
   }
-  clearTimeout(timeOut);
+
+  clearTimeout(debounceCallback.timeout);
   callbacks.delete(key);
   console.log(key, 'deleted');
 }
